combine changes in side-bar.svelte and +layout.svelte to side-bar-section
